+++ openwisp_controller/config/base/config.py
@@ -6,9 +6,10 @@ from collections import defaultdict
 from cache_memoize import cache_memoize
 from django.core.cache import cache
 from django.core.exceptions import ObjectDoesNotExist, PermissionDenied, ValidationError
+from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models, transaction
+from django.db.models import JSONField
 from django.utils.translation import gettext_lazy as _
-from jsonfield import JSONField
 from model_utils import Choices
 from model_utils.fields import StatusField
 from netjsonconfig import OpenWrt
@@ -85,13 +86,12 @@ class AbstractConfig(ChecksumCacheMixin, BaseConfig):
         blank=True,
         default=dict,
         help_text=_(
-            "Additional "
-            '<a href="http://netjsonconfig.openwisp.org/'
-            'en/stable/general/basics.html#context" target="_blank">'
-            "context (configuration variables)</a> in JSON format"
+            "allows overriding "
+            '<a href="https://openwisp.io/docs/stable/controller/user/variables.html'
+            '" target="_blank">'
+            "configuration variables</a>"
         ),
-        load_kwargs={"object_pairs_hook": collections.OrderedDict},
-        dump_kwargs={"indent": 4},
+        encoder=DjangoJSONEncoder,
     )
     checksum_db = models.CharField(
         _("configuration checksum"),