+++ openwisp_controller/config/signals.py
@@ -65,3 +65,11 @@ vpn_server_modified = Signal()
 vpn_server_modified.__doc__ = """
 providing arguments: ['instance']
 """
+whois_fetched = Signal()
+whois_fetched.__doc__ = """
+Providing arguments: ['whois', 'updated_fields', 'device']
+"""
+whois_lookup_skipped = Signal()
+whois_lookup_skipped.__doc__ = """
+Providing arguments: ['device']
+"""