+++ openwisp_controller/config/tests/test_api.py
@@ -29,8 +29,6 @@ VpnClient = load_model("config", "VpnClient")
 Device = load_model("config", "Device")
 Config = load_model("config", "Config")
 DeviceGroup = load_model("config", "DeviceGroup")
-DeviceLocation = load_model("geo", "DeviceLocation")
-Location = load_model("geo", "Location")
 OrganizationUser = load_model("openwisp_users", "OrganizationUser")
 
 
@@ -430,8 +428,8 @@ class TestConfigApi(
                 "backend": "netjsonconfig.OpenWisp",
                 "status": "modified",
                 "templates": [],
-                "context": "{}",
-                "config": "{}",
+                "context": {},
+                "config": {},
             },
         }
         r = self.client.put(path, data, content_type="application/json")