+++ CHANGES.rst
@@ -6,6 +6,42 @@ Version 1.3.0 [unreleased]
 
 Work in progress.
 
+Version 1.2.3 [2026-04-09]
+--------------------------
+
+Bugfixes
+~~~~~~~~
+
+- Corrected initial field value assignment in ``AbstractDevice`` class
+- Fixed duplicate template entries in Device admin
+- Invalidate Config checksum after subnet provisioning
+- Fixed concurrent ``update_config`` detection `#1204
+  <https://github.com/openwisp/openwisp-controller/issues/1204>`_
+
+Version 1.2.2 [2026-03-06]
+--------------------------
+
+Changes
+~~~~~~~
+
+Other changes
++++++++++++++
+
+- Improved help text of configuration variable fields
+- Minor fixes in the test suite
+
+Version 1.2.1 [2026-03-04]
+--------------------------
+
+Bugfixes
+~~~~~~~~
+
+- Use context variables in Vpn.auto_client for OpenVPN backend
+- Fixed 500 FieldError in DeviceLocationView `#1110
+  <https://github.com/openwisp/openwisp-controller/issues/1110>`_
+- Fixed MultiValueDictKeyError on empty device form submission `#1057
+  <https://github.com/openwisp/openwisp-controller/issues/1057>`_
+
 Version 1.2.0 [2025-10-24]
 --------------------------